\name{CRallow_alpha}
\alias{CRallow_alpha}
\title{
    Allows use of alpha channel.
}
\description{
    Enables use of the alpha channel in RGBA color arguments.
}
\usage{
CRallow_alpha(graphics_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer etc. created with \code{CRopen}.}
}
