\name{CRdraw_point}
\alias{CRdraw_point}
\title{
    Draws point.
}
\description{
    Draws a single point of pixel size to rendering target.
}
\usage{
CRdraw_point(graphics_list, point_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer etc. created with \code{CRopen}.}
    \item{point_list}{list defining single point, containing elements x and y.}
}
