\name{CRget_time}
\alias{CRget_time}
\title{
    Gives time in milliseconds.
}
\description{
    Get time in milliseconds since system was started.
}
\usage{
CRget_time(status = NULL)
}
\arguments{
    \item{status}{Optional list \code{status} with elements \code{value} and \code{message} to retrieve warnings and error messages.}
}
\value{
    Returns time in milliseconds since system was started.
}
