\name{CRhide_mouse}
\alias{CRhide_mouse}
\title{
    Hides mouse symbol.
}
\description{
    Hides mouse symbol.
}
\usage{
CRhide_mouse(status = NULL)
}
\arguments{
    \item{status}{Optional list \code{status} with elements \code{value} and \code{message} to retrieve warnings and error messages.}
}