\name{CRopen_controller}
\alias{CRopen_controller}
\title{
    Opens gamecontroller.
}
\description{
    Scans available joysticks, opens first gamecontroller among them, returns address of controller for eventual closing of device.
}
\usage{
CRopen_controller()
}
\value{
    Returns list containing pointer to gamecontroller.
}
