\name{CRreset_timer_resolution}
\alias{CreRset_timer_resolution}
\title{
    Reset global resolution of Windows timer.
}
\usage{
    CRreset_timer_resolution(resolution)
}
\arguments{
    \item{resolution}{the timer resolution previously set and returned by \code{CRset_timer_resolution}.}
}