\name{CRshow_mouse}
\alias{CRshow_mouse}
\title{
    Shows mouse again.
}
\description{
    Makes mouse symbol visible again after being hidden with \code{CRhide_mouse}.
}
\usage{
    CRshow_mouse(status = NULL)
}
\arguments{
    \item{status}{Optional list \code{status} with elements \code{value} and \code{message} to retrieve warnings and error messages.}
}