\name{CRwait_ms}
\alias{CRwait_ms}
\title{
    Waits a certain number of milliseconds.
}
\description{
    Waits a certain number of milliseconds.
}
\usage{
    CRwait_ms(ms, status = NULL)
}
\arguments{
    \item{ms}{Number of milliseconds to wait.}
    \item{status}{Optional list \code{status} with elements \code{value} and \code{message} to retrieve warnings and error messages.}
}