\name{CRdraw_rect}
\alias{CRdraw_rect}
\title{
    Draw a rectangle.
}
\description{
    Draws outline of rectangle defined in rect_list to rendering target.
}
\usage{
CRdraw_rect(graphics_list, rect_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer etc. created with \code{CRopen}.}
    \item{rect_list}{list defining a rectangle, containing elements x, y, w and h.}
}

