\name{CRfill_rect}
\alias{CRfill_rect}
\title{
    Draws a filled rectangle.
}
\description{
    Draws a filled rectangle defined in rect_list to rendering target.
}
\usage{
CRfill_rect(graphics_list, rect_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer etc. created with \code{CRopen}.}
    \item{rect_list}{list defining a rectangle, containing elements x, y, w and h.}
}

