\name{CRflush}
\alias{CRflush}
\title{
    Flushes event queue.
}
\description{
    Flushes all events in event queue specified in graphics_list.
}
\usage{
CRflush(graphics_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer etc. created with \code{CRopen}.}
}
