\name{CRget_time}
\alias{CRget_time}
\title{
    Get time in milliseconds.
}
\description{
    Get time in milliseconds since system was started.
}
\usage{
CRget_time(graphics_list)
}
\arguments{
    \item{graphics_list}{a list containing pointers to window, renderer, timer resolution, timer lag, etc., created with \code{CRopen}.}
}
\value{
    Returns time in milliseconds since system was started.
}

